refactor: enhance doc by adding @public API annotations and cleaning up comments
